Moved commands into packages
